# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a docs for TodoList.",
        "title": "TodoList API",
        "contact": {
            "name": "ChengXiao",
            "email": "cxxlxx0@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/group/{group}/task": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Get some group`s tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Create one task with group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group hashID",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task content",
                        "name": "task_content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/task": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Modify one task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task content",
                        "name": "task_content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "task deadline",
                        "name": "deadline",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "task status",
                        "name": "is_done",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "task hashID",
                        "name": "task",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Create one task without group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task content",
                        "name": "task_content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/task/default/": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get default`s tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/task/today/": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get Today`s tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/task/{task}": {
            "delete": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Delete one task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task hashID",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/user/password": {
            "patch": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change User Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "serializers.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}